requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai==0.3.1
python-dateutil==2.8.2